Update function for getting taxonomy tables
Add parameter for making the minimum 0.01% count threshold optional.
One use case is that alpha diversity measurements should be taken on unfiltered data

This uncovered a hidden problem: relying on the taxa name instead of lineage can lead to clashes when multiple taxa share a name (e.g. two genera from different lineages but with the same genus name).
When our genus data is not filtered, there are clashes, so I implemented a (kinda slow) fix that will append the next name in the lineage in brackets to (most-likely) solve the clash.
This solution is applied only to genus for now, since phylum are much less likely to share a name.