Refactor, add Woltka pathway analyses
Add formatting of the Process factor ANCOMBC2 results to match experimental design. This consists of flipping the direction of certain comparisons, and adjusting the signs of certain measurements, like log-fold change
Add reading in woltka pathway counts, and running ANCOMBC2 analyses on them as well.
Refactor and functionalize the main ancombc2.R script for a simpler main analysis loop